doc: update statistics documentation page
