+++ docs/api/_toctree/StatisticsAPI/miv.statistics.firing_rates.rst
@@ -1,0 +1,6 @@
+﻿miv.statistics.firing\_rates
+============================
+
+.. currentmodule:: miv.statistics
+
+.. autofunction:: firing_rates
