+++ docs/api/_toctree/StatisticsAPI/miv.statistics.inter_spike_intervals.rst
@@ -1,0 +1,6 @@
+﻿miv.statistics.inter\_spike\_intervals
+======================================
+
+.. currentmodule:: miv.statistics
+
+.. autofunction:: inter_spike_intervals
