--- docs/api/_toctree/StatisticsAPI/miv.statistics.spikestamps_statistics.rst
@@ -1,6 +1,0 @@
-﻿miv.statistics.spikestamps\_statistics
-======================================
-
-.. currentmodule:: miv.statistics
-
-.. autofunction:: spikestamps_statistics
